test long

+++ src/horizontal.py
@@ -125,7 +125,8 @@ for k, prompt_batch in enumerate(prompt_loader):
             if len(replay_buffer) == 0:
                 print(completions[0])
                 print(completions[1])
-
+            sequence_ids = sequence_ids.long()
+            returns = returns.long()
             
             rollout_indv.append(returns)
             returns = returns.to(device)
